feat(video): умный прогресс в модалке, иконки без текста, перенос скачивания в меню; fix(api): нормализация статусов/прогресса HeyGen; fix(db): миграции lesson_title и video_* для Postgres; fix(cache): duration=float, адаптивный HeyGen по умолчанию

+++ backend/services/video_cache_service.py
@@ -147,7 +147,7 @@ class VideoCacheService:
     
     def update_video_status(self, video_id: str, status: str, 
                            download_url: Optional[str] = None,
-                           duration: Optional[int] = None,
+                           duration: Optional[float] = None,
                            file_size: Optional[int] = None,
                            error_message: Optional[str] = None,
                            error_code: Optional[str] = None):
